% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beeca-package.R
\docType{package}
\name{beeca-package}
\alias{beeca}
\alias{beeca-package}
\title{beeca: Binary Endpoint Estimation with Covariate Adjustment}
\description{
Performs estimation of marginal treatment effects for binary outcomes when using logistic regression working models with covariate adjustment (see discussions in Magirr et al (2024) \url{https://osf.io/9mp58/}). Implements the variance estimators of Ge et al (2011) \doi{10.1177/009286151104500409} and Ye et al (2023) \doi{10.1080/24754269.2023.2205802}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openpharma.github.io/beeca/}
}

}
\author{
\strong{Maintainer}: Alex Przybylski \email{alexander.przybylski@novartis.com}

Authors:
\itemize{
  \item Mark Baillie \email{mark.baillie@novartis.com} (\href{https://orcid.org/0000-0002-5618-0667}{ORCID})
  \item Craig Wang \email{craig.wang@novartis.com} (\href{https://orcid.org/0000-0003-1804-2463}{ORCID})
  \item Dominic Magirr \email{dominic.magirr@novartis.com}
}

}
\keyword{internal}
