% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gt.R
\name{format_pvalue}
\alias{format_pvalue}
\title{Format p-value for clinical trial tables}
\usage{
format_pvalue(p, digits = 3, small_threshold = 0.001)
}
\arguments{
\item{p}{numeric p-value(s) to format}

\item{digits}{integer, number of decimal places. Default is 3.}

\item{small_threshold}{numeric, values below this threshold are displayed
as "<threshold". Default is 0.001.}
}
\value{
character vector of formatted p-values
}
\description{
Formats p-values according to common clinical trial reporting conventions.
Values less than 0.001 are displayed as "<0.001", otherwise rounded to
3 decimal places.
}
\examples{
format_pvalue(0.0001)
format_pvalue(c(0.05, 0.001, 0.0001))
}
