% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_model.geeglm}
\alias{sanitize_model.geeglm}
\title{(internal) Sanitize a geeglm model}
\usage{
\method{sanitize_model}{geeglm}(model, trt, ...)
}
\arguments{
\item{model}{a geeglm object from \link[geepack]{geeglm} with binomial family canonical link.}

\item{trt}{the name of the treatment variable on the right-hand side of the formula.}

\item{...}{ignored.}
}
\value{
if model is non-compliant will throw warnings or errors.
}
\description{
(internal) Sanitize a geeglm model
}
\keyword{internal}
