% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_model.glmgee}
\alias{sanitize_model.glmgee}
\title{(internal) Sanitize a glmgee model}
\usage{
\method{sanitize_model}{glmgee}(model, trt, ...)
}
\arguments{
\item{model}{a glmgee object from \link[glmtoolbox]{glmgee} with binomial family canonical link.}

\item{trt}{the name of the treatment variable on the right-hand side of the formula.}

\item{...}{ignored.}
}
\value{
if model is non-compliant will throw warnings or errors.
}
\description{
(internal) Sanitize a glmgee model
}
\keyword{internal}
