% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_variable}
\alias{sanitize_variable}
\title{(internal) Sanitize function to check model and data}
\usage{
sanitize_variable(model, trt)
}
\arguments{
\item{model}{an \link[stats]{glm} model object.}

\item{trt}{the name of the treatment variable on the right-hand side of the glm formula.}
}
\value{
if model and variable are non-compliant, will throw warnings or error.
}
\description{
(internal) Sanitize function to check model and data
}
\keyword{internal}
